Extract enterprise authorization policy module
